fix(ports): drop label cache entries when workspaces are deleted
Previously the label cache held a workspaceId → labels map for the life
of the process. After a workspace was deleted, the next lookup would
still return cached labels (or the cached "null" sentinel), both of
which are stale.

Desktop: split the cache out of the ports router so workspace-delete
paths can invalidate without creating a ports ↔ workspaces import cycle,
then call it from `deleteWorkspace` and the project-cascade delete.

Host-service: call `invalidateLabelCache` from `workspace.delete` and
`workspace-cleanup` after the sqlite row is removed.

Rollback-on-create paths are intentionally skipped — the workspace never
had a cache entry to begin with.
