+++ apps/desktop/src/lib/trpc/routers/projects/projects.ts
@@ -26,6 +26,7 @@ import { PROJECT_COLOR_VALUES } from "shared/constants/project-colors";
 import { z } from "zod";
 import { publicProcedure, router } from "../..";
 import { resolveDefaultEditor } from "../external";
+import { invalidatePortLabelCache } from "../ports/label-cache";
 import {
 	activateProject,
 	getBranchWorkspace,
@@ -1577,6 +1578,7 @@ export const createProjectsRouter = (getWindow: () => BrowserWindow | null) => {
 						.delete(workspaces)
 						.where(inArray(workspaces.id, closedWorkspaceIds))
 						.run();
+					for (const id of closedWorkspaceIds) invalidatePortLabelCache(id);
 				}
 
 				localDb
