+++ apps/desktop/src/lib/trpc/routers/workspaces/utils/db-helpers.ts
@@ -11,6 +11,7 @@ import {
 import { and, desc, eq, isNotNull, isNull } from "drizzle-orm";
 
 import { localDb } from "main/lib/local-db";
+import { invalidatePortLabelCache } from "../../ports/label-cache";
 import { computeNextProjectChildTabOrder } from "./project-children-order";
 
 /**
@@ -273,6 +274,7 @@ export function clearWorkspaceDeletingStatus(workspaceId: string): void {
  */
 export function deleteWorkspace(workspaceId: string): void {
 	localDb.delete(workspaces).where(eq(workspaces.id, workspaceId)).run();
+	invalidatePortLabelCache(workspaceId);
 }
 
 /**
